refactor: 주문 테이블의 서비스네이밍을 변경하고 Order 상태에 대한 서비스네이밍을 명시적으로 수정하라
