+++ src/test/java/kitchenpos/table/application/OrderTableServiceTest.java
@@ -17,7 +17,7 @@ import kitchenpos.common.exception.ExceptionType;
 import kitchenpos.common.exception.NotFoundException;
 import kitchenpos.order.domain.OrderRepository;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.application.OrderTableServiceImpl;
+import kitchenpos.order.application.OrderTableStatusServiceImpl;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.OrderTableRepository;
 import kitchenpos.table.dto.request.OrderTableRequest;
@@ -32,16 +32,16 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 @DisplayName("테이블 서비스에 대한 테스트")
 @ExtendWith(MockitoExtension.class)
-class TableServiceTest {
+class OrderTableServiceTest {
 
     @Mock
     private OrderTableRepository orderTableRepository;
     @Mock
-    private OrderTableService orderTableService;
+    private OrderTableStatusService orderTableStatusService;
 
 
     @InjectMocks
-    private TableService tableService;
+    private OrderTableService orderTableService;
 
     private OrderTableRequest 주문_테이블_request;
     private OrderTableRequest 주문_테이블_2_request;
@@ -64,7 +64,7 @@ class TableServiceTest {
             .thenReturn(주문_테이블);
 
         // when
-        OrderTableResponse result = tableService.create(주문_테이블_request);
+        OrderTableResponse result = orderTableService.create(주문_테이블_request);
 
         // then
         assertThat(result).isNotNull();
@@ -78,7 +78,7 @@ class TableServiceTest {
             .thenReturn(Arrays.asList(주문_테이블, 주문_테이블));
 
         // when
-        List<OrderTableResponse> result = tableService.list();
+        List<OrderTableResponse> result = orderTableService.list();
 
         // then
         assertThat(result).hasSize(2);
@@ -91,13 +91,13 @@ class TableServiceTest {
         주문_테이블.fillTheTable();
         when(orderTableRepository.findById(주문_테이블.getId()))
             .thenReturn(Optional.of(주문_테이블));
-        doNothing().when(orderTableService)
+        doNothing().when(orderTableStatusService)
             .validateOrderTableStatus(주문_테이블);
         when(orderTableRepository.save(주문_테이블))
             .thenReturn(주문_테이블);
 
         // when
-        OrderTableResponse result = tableService.changeEmpty(주문_테이블.getId());
+        OrderTableResponse result = orderTableService.changeEmpty(주문_테이블.getId());
 
         // then
         assertTrue(result.isEmpty());
@@ -112,7 +112,7 @@ class TableServiceTest {
 
         // then
         assertThatThrownBy(() -> {
-            tableService.changeEmpty(주문_테이블.getId());
+            orderTableService.changeEmpty(주문_테이블.getId());
         }).isInstanceOf(NotFoundException.class)
             .hasMessageContaining(ExceptionType.NOT_EXIST_ORDER_TABLE.getMessage());
     }
@@ -122,18 +122,18 @@ class TableServiceTest {
     void change_empty_exception_test2() {
         // given
         OrderRepository orderRepository = mock(OrderRepository.class);
-        OrderTableService orderTableService = new OrderTableServiceImpl(orderRepository);
+        OrderTableStatusService orderTableStatusService = new OrderTableStatusServiceImpl(orderRepository);
         when(orderTableRepository.findById(주문_테이블.getId()))
             .thenReturn(Optional.of(주문_테이블));
         when(orderRepository.existsByOrderTableIdAndOrderStatusIn(주문_테이블.getId(), Arrays.asList(
             OrderStatus.COOKING, OrderStatus.MEAL)))
             .thenReturn(true);
 
-        TableService tableService = new TableService(orderTableRepository, orderTableService);
+        OrderTableService orderTableService = new OrderTableService(orderTableRepository, orderTableStatusService);
 
         // then
         assertThatThrownBy(() -> {
-            tableService.changeEmpty(주문_테이블.getId());
+            orderTableService.changeEmpty(주문_테이블.getId());
         }).isInstanceOf(CannotUpdateException.class)
             .hasMessageContaining(ExceptionType.CAN_NOT_UPDATE_TABLE_IN_COOKING_AND_MEAL_STATUS.getMessage());
     }
@@ -149,7 +149,7 @@ class TableServiceTest {
             .thenReturn(주문_테이블);
 
         // when
-        OrderTableResponse result = tableService.changeNumberOfGuests(주문_테이블.getId(), 주문_테이블_2_request);
+        OrderTableResponse result = orderTableService.changeNumberOfGuests(주문_테이블.getId(), 주문_테이블_2_request);
 
         // then
         assertThat(result.getNumberOfGuests()).isEqualTo(주문_테이블_2_request.getNumberOfGuests());
@@ -163,7 +163,7 @@ class TableServiceTest {
 
         // then
         assertThatThrownBy(() -> {
-            tableService.changeNumberOfGuests(주문_테이블.getId(), 주문_테이블_request);
+            orderTableService.changeNumberOfGuests(주문_테이블.getId(), 주문_테이블_request);
         }).isInstanceOf(BadRequestException.class)
             .hasMessageContaining(ExceptionType.CAN_NOT_LESS_THAN_ZERO_GUESTS.getMessage());
     }
@@ -177,7 +177,7 @@ class TableServiceTest {
 
         // then
         assertThatThrownBy(() -> {
-            tableService.changeNumberOfGuests(주문_테이블.getId(), 주문_테이블_request);
+            orderTableService.changeNumberOfGuests(주문_테이블.getId(), 주문_테이블_request);
         }).isInstanceOf(NotFoundException.class)
             .hasMessageContaining(ExceptionType.NOT_EXIST_ORDER_TABLE.getMessage());
     }
@@ -191,7 +191,7 @@ class TableServiceTest {
 
         // then
         assertThatThrownBy(() -> {
-            tableService.changeNumberOfGuests(주문_테이블.getId(), 주문_테이블_request);
+            orderTableService.changeNumberOfGuests(주문_테이블.getId(), 주문_테이블_request);
         }).isInstanceOf(BadRequestException.class)
             .hasMessageContaining(ExceptionType.EMPTY_TABLE.getMessage());
     }
